# openapi.yaml
openapi: '3.0.3'
info:
  title: Census App API
  description: REST API for the Census App — a household and record management system.
  version: '1.0.0'
servers:
  - url: http://localhost:3000
    description: Localhost
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in with credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Authenticated }
        '401': { description: Invalid credentials }
  /api/auth/token:
    post:
      tags: [Auth]
      summary: Create bearer token for API clients (Swagger)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Bearer token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  tokenType: { type: string, example: Bearer }
                  expiresIn: { type: integer, example: 7200 }
        '400': { description: Invalid request payload }
        '401': { description: Invalid credentials }
        '500': { description: Server configuration error }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...